+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 - Change backup compression to the more effective "Zstandard" algorithm
 - Change to accept ssh host keys for new hosts
 - Change to no longer support migration from Pika Backup v0.2 configs
+- Add better error messages for incorrect remote locations
 - Add message to user interface if target devive or server is unrepsonive
+- Add support for user@host:path borg remote location syntax
 
 ## 0.3.5 (2021-09-08)